Fix & simplify names. * Rename package step to task * Simplify RunTestCorelliumAndroid to TestAndroid * Simplify RunTestCorelliumAndroidCommand to TestAndroidCommand
