--- corellium/cli/src/main/kotlin/flank/corellium/cli/run/test/android/task/JUnit.kt
@@ -1,7 +1,0 @@
-package flank.corellium.cli.run.test.android.task
-
-import flank.exection.parallel.type
-import flank.exection.parallel.using
-import flank.junit.JUnit
-
-internal val jUnitApi = type<JUnit.Api>() using { JUnit.Api() }
