+++ corellium/cli/src/main/kotlin/flank/corellium/cli/test/android/task/Apk.kt
@@ -1,0 +1,7 @@
+package flank.corellium.cli.test.android.task
+
+import flank.apk.Apk
+import flank.exection.parallel.type
+import flank.exection.parallel.using
+
+internal val apkApi = type<Apk.Api>() using { Apk.Api() }
