+++ corellium/cli/src/main/kotlin/flank/corellium/cli/test/android/task/JUnit.kt
@@ -1,0 +1,7 @@
+package flank.corellium.cli.test.android.task
+
+import flank.exection.parallel.type
+import flank.exection.parallel.using
+import flank.junit.JUnit
+
+internal val jUnitApi = type<JUnit.Api>() using { JUnit.Api() }
