+++ corellium/cli/src/test/kotlin/flank/corellium/cli/RunTestCorelliumAndroidCommandTest.kt
@@ -1,6 +1,6 @@
 package flank.corellium.cli
 
-import flank.corellium.cli.RunTestCorelliumAndroidCommand.Companion.resolve
+import flank.corellium.cli.TestAndroidCommand.Companion.resolve
 import flank.exection.parallel.ParallelState
 import flank.exection.parallel.validate
 import org.junit.Test
@@ -10,7 +10,7 @@ class RunTestCorelliumAndroidCommandTest {
     @Test
     fun validate() {
         val initial: ParallelState = mapOf(
-            RunTestCorelliumAndroidCommand to RunTestCorelliumAndroidCommand()
+            TestAndroidCommand to TestAndroidCommand()
         )
 
         resolve.validate(initial)
