+++ corellium/domain/src/main/kotlin/flank/corellium/domain/test/android/task/Finish.kt
@@ -1,0 +1,15 @@
+package flank.corellium.domain.test.android.task
+
+import flank.corellium.domain.TestAndroid.CompleteTests
+import flank.corellium.domain.TestAndroid.DumpShards
+import flank.corellium.domain.TestAndroid.GenerateReport
+import flank.exection.parallel.from
+import flank.exection.parallel.using
+
+/**
+ * The final step, notifies that execution completes without exceptions.
+ */
+internal val finish = CompleteTests from setOf(
+    GenerateReport,
+    DumpShards
+) using { }
